Change directory configuration
Base.scala Parser.scala Gungnir.scala to src/Gungnir/

Fix compile.bat
Fix README.md
