+++ src/Gungnir/Proposition.gung
